auth and powers of each users plus a test user inside db:seed

+++ app/Http/Controllers/JobController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -11,6 +12,7 @@ use Illuminate\View\View;
 // php artisan make:controller JobController --resource
 class JobController extends Controller
 {
+    use AuthorizesRequests;
 
     /**
      * @desc Display a listing of jobs.
@@ -30,7 +32,7 @@ class JobController extends Controller
      * @method GET
      * @return View
      */
-    public function create(): View
+    public function create(): mixed
     {
         return view('jobs.create');
     }
@@ -67,7 +69,7 @@ class JobController extends Controller
             'company_website' => 'nullable|url',
         ]);
 
-        $validateData['user_id'] = 1;
+        $validateData['user_id'] = auth()->user()->id;
 
         if($request->hasFile('company_logo')){
             $path = $request->file('company_logo')->store('logos', 'public');
@@ -100,6 +102,7 @@ class JobController extends Controller
      */
     public function edit(Job $job): View
     {
+        $this->authorize('update', $job);
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -113,6 +116,8 @@ class JobController extends Controller
      */
     public function update(Request $request, Job $job): string
     {
+        //check if user is authoridez
+        $this->authorize('update', $job);
         $validateData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -153,6 +158,7 @@ class JobController extends Controller
      */
     public function destroy(Job $job): RedirectResponse
     {
+        $this->authorize('delete', $job);
         if($job->company_logo){
             Storage::delete('public/logos/', $job->company_logo);
         }
